the Level Engine is here
A single object that will handle UI, scene loading, win & lose conditions, and more. Only thing needed to be modified each level is what the NextScene to load is in the EventHandler, and adjusting the timer on each level. This will automatically count how many WinPlatforms and Players are in each scene and adjust the win conditions without any need to editing.